feat(tool, lib): Add configurable cleanup method

+++ TypeContractor/Contractor.cs
@@ -39,7 +39,7 @@ public class Contractor
 
     public TypeContractorConfiguration Configuration => _configuration;
 
-    public void Build(MetadataLoadContext? metadataLoadContext = null)
+    public void Build(MetadataLoadContext? metadataLoadContext = null, bool smartClean = false)
     {
         metadataLoadContext ??= BuildMetadataLoadContext();
         var toConvert = new List<ContractedType>();
@@ -84,41 +84,44 @@ public class Contractor
             }
         }
 
-        _configuration.Logger.LogMessage("Cleaning no longer relevant output files.");
-        var allFiles = Directory.GetFiles(_configuration.OutputPath, "*", SearchOption.AllDirectories);
-        var diff = allFiles.Except(generatedFiles).ToList();
-
-        var allDirectories = Directory.GetDirectories(_configuration.OutputPath, "*", SearchOption.AllDirectories);
-        var generatedDirectories = allDirectories.Where(d => generatedFiles.Any(f => f.StartsWith(d, StringComparison.InvariantCultureIgnoreCase)));
-        var directoryDiff = allDirectories.Except(generatedDirectories).ToList();
-
-        _configuration.Logger.LogDebug($"Found a diff of {diff.Count} items:");
-        foreach (var file in diff)
+        if (smartClean)
         {
-            _configuration.Logger.LogDebug($"  {file}");
-            try
+            _configuration.Logger.LogMessage("Cleaning no longer relevant output files.");
+            var allFiles = Directory.GetFiles(_configuration.OutputPath, "*", SearchOption.AllDirectories);
+            var diff = allFiles.Except(generatedFiles).ToList();
+
+            var allDirectories = Directory.GetDirectories(_configuration.OutputPath, "*", SearchOption.AllDirectories);
+            var generatedDirectories = allDirectories.Where(d => generatedFiles.Any(f => f.StartsWith(d, StringComparison.InvariantCultureIgnoreCase)));
+            var directoryDiff = allDirectories.Except(generatedDirectories).ToList();
+
+            _configuration.Logger.LogDebug($"Found a diff of {diff.Count} items:");
+            foreach (var file in diff)
             {
-            if (File.Exists(file))
-                File.Delete(file);
-        }
-            catch (IOException ex)
-            {
-                _configuration.Logger.LogWarning($"Unable to delete file {file}: {ex.Message}");
+                _configuration.Logger.LogDebug($"  {file}");
+                try
+                {
+                    if (File.Exists(file))
+                        File.Delete(file);
+                }
+                catch (IOException ex)
+                {
+                    _configuration.Logger.LogWarning($"Unable to delete file {file}: {ex.Message}");
+                }
             }
-        }
 
-        _configuration.Logger.LogDebug($"Found a diff of {directoryDiff.Count} folders:");
-        foreach (var dir in directoryDiff)
-        {
-            _configuration.Logger.LogDebug($"  {dir}");
-            try
+            _configuration.Logger.LogDebug($"Found a diff of {directoryDiff.Count} folders:");
+            foreach (var dir in directoryDiff)
             {
-            if (Directory.Exists(dir))
-                Directory.Delete(dir, true);
-        }
-            catch (IOException ex)
-            {
-                _configuration.Logger.LogWarning($"Unable to delete directory {dir}: {ex.Message}");
+                _configuration.Logger.LogDebug($"  {dir}");
+                try
+                {
+                    if (Directory.Exists(dir))
+                        Directory.Delete(dir, true);
+                }
+                catch (IOException ex)
+                {
+                    _configuration.Logger.LogWarning($"Unable to delete directory {dir}: {ex.Message}");
+                }
             }
         }
     }
